pyproject.toml
[project]
name = "tanaka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.0",
    "hikari[speedups]>=2.3.3",
    "uvloop>=0.21.0",
    "hikari-lightbulb>=3.0.1",
    "aiozipstream>=0.4",
]

[dependency-groups]
dev = [
    "nox>=2025.5.1",
    "ruff>=0.11.13",
]
